Fix NSF external channel mixing. Remove old PPU colorChart filling code. Add NTSC Scaler. Add support for using internal NTSC simulated palette. Fix DirectX10 renderer message scaling issues. Optimize nearest neighbor scaler.

+++ DirectXEmu/EmuoTron/Channels/NSF.cs
@@ -171,7 +171,7 @@ namespace EmuoTron.Channels
 
         public override byte Cycle()
         {
-            byte volume = 0;
+            int volume = 0;
             volume += VRC6.Cycle();
             volume += VRC7.Cycle();
             volume += FDS.Cycle();
@@ -183,7 +183,7 @@ namespace EmuoTron.Channels
             {
                 nes.mapper.interruptMapper = ((Channels.MMC5)MMC5).interrupt;
             }
-            return volume;
+            return (byte)volume;
         }
 
         public override void StateSave(System.IO.BinaryWriter writer)
